requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
cryptography==2.7
future==0.17.1
pipenv==2018.11.26
pycparser==2.19
python-telegram-bot==12.0.0b1
six==1.12.0
tornado==6.0.2
virtualenv==16.6.0
virtualenv-clone==0.5.3
imgurpython==1.1.7